Removed DesktopWallpaperHandler from FullTrust. Added SetSlideShow method to WallpaperHelpers.
